pyproject.toml
[project]
name = "voice-recorder"
version = "0.1.0"
description = "Voice recorder with VAD"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.build.targets.wheel]
include = ["server.py"]